% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_helpers.R
\name{as.numeric.factor}
\alias{as.numeric.factor}
\title{Convert factors to numerics (non-numeric values are converted to NA)}
\usage{
\method{as.numeric}{factor}(x)
}
\arguments{
\item{x}{factor vector}
}
\value{
numerics vector
}
\description{
Convert factors to numerics (non-numeric values are converted to NA)
}
