% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_helpers.R
\name{average_within_and_between_sites}
\alias{average_within_and_between_sites}
\title{Average values first from same sites and then between sites}
\usage{
average_within_and_between_sites(data, site_names)
}
\arguments{
\item{data}{Vector with values}

\item{site_names}{Site names of each vector element}
}
\value{
Average values
}
\description{
Average values first from same sites and then between sites
}
