% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_helpers.R
\name{bayesgamlsinterp}
\alias{bayesgamlsinterp}
\title{Fit Bayesian GAMLS with `brms::brm` and return multivariate zoo with the samples for the location and scales parameters in the columns}
\usage{
bayesgamlsinterp(
  xin,
  xout,
  familyname = "beta",
  smooth_scale = NULL,
  location = TRUE,
  scale = TRUE,
  nr_samples = 1000,
  backend = "rstan"
)
}
\arguments{
\item{xin}{Input zoo}

\item{xout}{Output time axis}

\item{familyname}{Response distribution family (note that unlike gaminterp, here a string with a family name is required not the actually family function, "beta" is the only valid option at the moment); this is because different families have different parameter names to parameterize the gamls model}

\item{smooth_scale}{Approximate smoothing scale of the GAM model (translated internally to number of basis functions of the fitted splines)}

\item{location}{Should samples of location parameter be returned?}

\item{scale}{Should samples of scale parameter be returned?}

\item{nr_samples}{Number of posterior samples for each parameter}

\item{backend}{Which stan backend should be used ("rstan" or "cmdstanr")}
}
\value{
Zoo with samples from location (smoothed and interpolated values) and scale (magnitude of residuals, measure for the approximate magnitude of sub-smoothscale variability) at the provided interpolation dates
}
\description{
Fit Bayesian GAMLS with `brms::brm` and return multivariate zoo with the samples for the location and scales parameters in the columns
}
