% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_helpers.R
\name{binning}
\alias{binning}
\title{Binning}
\usage{
binning(xin, xout, bin_width = mean(diff(xout)), binning_function = mean)
}
\arguments{
\item{xin}{Input zoo}

\item{xout}{Output time axis}

\item{bin_width}{Width of bins if interpolation method is "binning". Defaults to the mean sample resolution (no variable bin sizes are supported at the moment)}

\item{binning_function}{How should values within one bin be averaged? Default is "mean"}
}
\value{
Zoo with binned values at the provided interpolation dates
}
\description{
Binning
}
