% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_helpers.R
\name{clean_timeseries}
\alias{clean_timeseries}
\title{Remove NAs and aggregate non-unique timesteps in zoo objects}
\usage{
clean_timeseries(
  xin,
  remove_na = TRUE,
  aggregation = TRUE,
  aggregation_fun = mean
)
}
\arguments{
\item{xin}{Proxytibble with proxy data in `zoo::zoo` format, or irregular time series object (`zoo::zoo`), xin can be multivariate}

\item{remove_na}{Flag if NAs should be removed}

\item{aggregation}{Flag if non-unique timesteps should be merged}

\item{aggregation_fun}{Function for merging non-unique timesteps}
}
\value{
Proxytibble with proxy data in `zoo::zoo` format, or irregular time series object (`zoo::zoo`)
}
\description{
Remove NAs and aggregate non-unique timesteps in zoo objects
}
\examples{
# Load Monticchio example data from PTBoxProxydata
library(PTBoxProxydata)
mng <- PTBoxProxydata::ProxyDataManager()
monticchiodata <- PTBoxProxydata::load_set(mng,'monticchio_testset',zoo_format = 'zoo', force_file=TRUE)
# Clean proxy data time series
monticchiodata_cleaned <- clean_timeseries(monticchiodata)
# Clean zoo time series
monticchiozoo_cleaned <- clean_timeseries(monticchiodata$proxy_data[[1]])
# Plot zoo data
plot(monticchiozoo_cleaned)
plot(monticchiodata_cleaned$proxy_data[[1]])

}
\seealso{
\link{aggregate} (from `stats`) for merging non-unique timesteps
}
