% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_helpers.R
\name{compute_group_weights_from_maps}
\alias{compute_group_weights_from_maps}
\title{Computes spatial weights for groups based on the relative area covered by them}
\usage{
compute_group_weights_from_maps(group_maps)
}
\arguments{
\item{group_maps}{Input list with maps for each group}
}
\value{
Named vector with weights for all groups (they only sum up to 1 if all NA grid boxes are filled by exactly one group)
}
\description{
Computes spatial weights for groups based on the relative area covered by them
}
