% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_helpers.R
\name{cor_test_effdof}
\alias{cor_test_effdof}
\title{Compute Pearson / Spearman correlations with adjusted estimates for CIs / p-values that use effective degrees of freedom (computed from autocorrelation estimates with nest) instead of iid assumption
Useful for computing correlations of auto-correlated timeseries that share the same time axis}
\usage{
cor_test_effdof(
  xin,
  yin,
  alternative = c("two.sided", "less", "greater"),
  method = c("pearson", "spearman"),
  conf.level = 0.95,
  dt = NULL,
  continuity = FALSE
)
}
\arguments{
\item{xin}{Input vector}

\item{yin}{Input vector}

\item{alternative}{Type of null hypothesis}

\item{method}{Pearson of spearman correlation}

\item{conf.level}{confidence level for the returned confidence interval. Currently only used for the Pearson product moment correlation coefficient if there are at least 4 complete pairs of observations.}

\item{dt}{Sampling rate for estimating persistence of the timeseries}

\item{continuity}{logical: if true, a continuity correction is used for and Spearman correlation when not computed exactly.}
}
\value{
list with same parameters as `stats::cor.test`
}
\description{
Compute Pearson / Spearman correlations with adjusted estimates for CIs / p-values that use effective degrees of freedom (computed from autocorrelation estimates with nest) instead of iid assumption
Useful for computing correlations of auto-correlated timeseries that share the same time axis
}
\seealso{
\link{cor.test} (from `stats`) for version with independent samples
\link{nexcf} (from `nest`) for non-pairwise correlation estimation method which uses the same formula to estimate the dof
}
