% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_helpers.R
\name{gaminterp}
\alias{gaminterp}
\title{GAM interpolation and smoothing}
\usage{
gaminterp(xin, xout, family = gaussian, smooth_scale = NULL)
}
\arguments{
\item{xin}{Input zoo}

\item{xout}{Output time axis}

\item{family}{Response distribution family}

\item{smooth_scale}{Approximate smoothing scale of the GAM model (translated internally to number of basis functions of the fitted splines)}
}
\value{
Zoo with smoothed and interpolated values at the provided interpolation dates
}
\description{
GAM interpolation and smoothing
}
