% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_tools.R
\name{get_var_from_spec_wrapper}
\alias{get_var_from_spec_wrapper}
\title{Wrapper to compute frequency-dependent variance (based on `PaleoSpec::GetVarFromSpectra`)
DEPRICATED --> Use `tsc_and_period_dep_var` instead, which provides point estimates and confidence intervals}
\usage{
get_var_from_spec_wrapper(xin, freq.start, freq.end, detrend = FALSE)
}
\arguments{
\item{xin}{Input zoo}

\item{freq.start}{Lower frequency limit}

\item{freq.end}{Upper frequency limit}

\item{detrend}{Should linear detrending be applied before computing the spectra from which the timescales dependent variance is deduced}
}
\value{
Vector with variance and degrees of freedom estimates
}
\description{
Wrapper to compute frequency-dependent variance (based on `PaleoSpec::GetVarFromSpectra`)
DEPRICATED --> Use `tsc_and_period_dep_var` instead, which provides point estimates and confidence intervals
}
