% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_tools.R
\name{get_var_from_spec_wrapper_logsmooth}
\alias{get_var_from_spec_wrapper_logsmooth}
\title{Wrapper to compute frequency-dependent variance from logsmoothed spectra (based on `PaleoSpec::GetVarFromSpectra`)
SHOULD NOT BE USED FOR COMPUTING dof, ONLY var ESTIMATES SEEM REASONABLE
DEPRICATED --> Use `tsc_and_period_dep_var` instead, which provides point estimates and confidence intervals}
\usage{
get_var_from_spec_wrapper_logsmooth(
  xin,
  freq.start,
  freq.end,
  detrend = FALSE,
  df_log = 0.05,
  bLog = FALSE
)
}
\arguments{
\item{xin}{Input zoo}

\item{freq.start}{Lower frequency limit}

\item{freq.end}{Upper frequency limit}

\item{detrend}{Should linear detrending be applied before computing the spectra from which the timescales dependent variance is deduced}

\item{df_log}{Width of the log-smoother in log units}

\item{bLog}{TRUE: average in the log space of the power, FALSE: arithmetic average}
}
\value{
Vector with variance and degrees of freedom estimates
}
\description{
Wrapper to compute frequency-dependent variance from logsmoothed spectra (based on `PaleoSpec::GetVarFromSpectra`)
SHOULD NOT BE USED FOR COMPUTING dof, ONLY var ESTIMATES SEEM REASONABLE
DEPRICATED --> Use `tsc_and_period_dep_var` instead, which provides point estimates and confidence intervals
}
