% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_helpers.R
\name{gkinterp}
\alias{gkinterp}
\title{Gaussian kernel interpolation and smoothing}
\usage{
gkinterp(xin, xout, smooth_scale, pass = 0.5)
}
\arguments{
\item{xin}{Input zoo}

\item{xout}{Output time axis}

\item{smooth_scale}{Smoothing scale of the Gaussian kernel}

\item{pass}{Gain at the smoothing scale}
}
\value{
Zoo with smoothed and interpolated values at the provided interpolation dates
}
\description{
Gaussian kernel interpolation and smoothing
}
\seealso{
\link{ksmooth} (from `stats`) is used for the smoothing and interpolation, but with modified definition of the kernel width to align better with timescale point of view
}
