% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_helpers.R
\name{interp_bwr25}
\alias{interp_bwr25}
\title{Anti-aliasing linear interpolation method from Baudouin et al. 2025 ("bwr25")}
\usage{
interp_bwr25(xin, xout)
}
\arguments{
\item{xin}{Input zoo}

\item{xout}{Output time axis (only equally spaced xout are supported at the moment)}
}
\value{
Zoo with interpolated values
}
\description{
Anti-aliasing linear interpolation method from Baudouin et al. 2025 ("bwr25")
}
