% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_tools.R
\name{multi_tsc_and_period_dep_var}
\alias{multi_tsc_and_period_dep_var}
\title{Wrapper to estimate timescale and time period dependent variances for multiple periods and timescale ranges}
\usage{
multi_tsc_and_period_dep_var(
  xin,
  period_starts,
  period_ends,
  tsc_lows,
  tsc_ups
)
}
\arguments{
\item{xin}{Input zoo}

\item{period_starts}{Start of periods of interest}

\item{period_ends}{End of periods of interest}

\item{tsc_lows}{Upper filtering timescales}

\item{tsc_ups}{Lower filtering timescales}

\item{detrend}{Should linear detrending be applied before computing the spectra from which the timescales dependent variance is deduced}
}
\value{
3D or 4D Matrix (Input zoo dimension x number of periods x number of timescales x output parameters); output parameters are the variance estimate, estimated dof, and lower and upper bounds of the confidence intervals
}
\description{
Wrapper to estimate timescale and time period dependent variances for multiple periods and timescale ranges
}
