% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_tools.R
\name{paleodata_signal_extraction}
\alias{paleodata_signal_extraction}
\title{Extract low-dimensional signals from multivariate datasets (e.g. pollen assemblages)}
\usage{
paleodata_signal_extraction(xin, signal_type, signal_components = 1)
}
\arguments{
\item{xin}{Proxytibble with multivariate proxy data in `zoo::zoo` format, or a multivariate irregular time series object (`zoo::zoo`)}

\item{signal_type}{Method to extract signals;
Implemented methods:
'pca' Principal component analysis
'dca' Detrended correspondance analysis
'ca' Correspondance analysis
'prc' Principal curves}

\item{signal_components}{Components that should be returned (e.g. c(2,3) for 'pca' returns the second and third principal components)}
}
\value{
Proxytibble with extracted signals in proxy data, or irregular time series object (`zoo::zoo`) containing the extracted signals
}
\description{
Extract low-dimensional signals from multivariate datasets (e.g. pollen assemblages)
}
\examples{
# Load ice core example data
library(PTBoxProxydata)
mng <- ProxyDataManager()
icecoredata <- load_set(mng,'icecore_testset',zoo_format = 'zoo')
# Compute PCA from multivariate zoo's
icecoredata_pca <- paleodata_signal_extraction(icecoredata, 'pca')
plot(icecoredata_pca$proxy_data[[1]])
plot(icecoredata_pca$proxy_data[[2]])
# Compute principal curve from multivariate zoo's
icecoredata_prc <- paleodata_signal_extraction(icecoredata, 'prc')
plot(icecoredata_prc$proxy_data[[1]])
plot(icecoredata_prc$proxy_data[[2]])

}
\seealso{
\link{prcomp} (from `stats`) for principal component analysis

\link{decorana} (from `vegan`) for detrended correspondance analysis

\link{cca} (from `vegan`) for correspondance analysis

\link{principal_curve} (from `princurve`) for principal curves
}
