% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_tools.R
\name{paleodata_spectrum}
\alias{paleodata_spectrum}
\title{Function to compute spectra (raw and smoothed using log-smooth) following the multi-taper methodology applied in Laepple and Huybers 2014, Rehfeld et al. 2018}
\usage{
paleodata_spectrum(
  xin,
  interpolation = TRUE,
  interpolation_type = "spectral",
  interpolation_dates = NULL,
  transformation = FALSE,
  transformation_type = "normalize",
  detrend = TRUE,
  df_log = 0.05,
  bLog = FALSE
)
}
\arguments{
\item{xin}{Proxytibble with proxy data in `zoo::zoo` format, or irregular time series object (`zoo::zoo`), xin can be multivariate}

\item{interpolation}{Logical: should interpolation be applied?}

\item{interpolation_type}{Parameters for interpolation (see paleodata_interpolation)}

\item{interpolation_dates}{Parameters for interpolation (see paleodata_interpolation). If NULL the interpolation is set to the mean temporal resolution of each proxy time series.}

\item{transformation}{Logical: should transformation be applied?}

\item{transformation_type}{Parameters for transformation (see paleodata_transformation)}

\item{detrend}{Logical: should time series be linearly detrend before computing the spectrum}

\item{df_log}{Width of the log-smoother in log units}

\item{bLog}{TRUE: average in the log space of the power, FALSE: arithmetic average}
}
\value{
List with raw and log-smoothed spectra (object description in SpecMTM and LogSmooth functions of PaleoSpec package)
}
\description{
Function to compute spectra (raw and smoothed using log-smooth) following the multi-taper methodology applied in Laepple and Huybers 2014, Rehfeld et al. 2018
}
\examples{
# Load ice core example data
library(PTBoxProxydata)
library(PaleoSpec)
mng <- ProxyDataManager()
icecoredata <- load_set(mng,'icecore_testset',zoo_format = 'zoo')
# First compute spectra
testout <- icecoredata$proxy_data[[1]] \%>\% paleodata_spectrum(.)
# Second plot spectra (output is list with raw and log-smoothed spectra that can be plotted using PaleoSpec())
PaleoSpec::LPlot(testout$raw)
PaleoSpec::LLines(testout$logsmooth)

}
\seealso{
\link{paleodata_interpolation} (from `PTBoxProxytools`) for interpolation to regular time axis

\link{paleodata_transformation} (from `PTBoxProxytools`) for time series transformation

\link{SpecMTM} (from `PaleoSpec`) for computation of spectrum with multi-taper methodology

\link{LogSmooth} (from `PaleoSpec`) for log smoothing of spectrum
}
