% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_tools.R
\name{paleodata_transformation}
\alias{paleodata_transformation}
\title{Transform data in an irregular time series object}
\usage{
paleodata_transformation(xin, transformation_type = "logit")
}
\arguments{
\item{xin}{Proxytibble with proxy data in `zoo::zoo` format, or irregular time series object (`zoo::zoo`), xin can be multivariate}

\item{transformation_type}{Type of transformation;
Implemented methods:
'logit' (logit transformation),
invlogit' (inverse logit transformation),
'probit' (probit transformation),
'invprobit' (inverse probit transformation),
'sqrt' (square-root transformation)
'quadratic' (quadratic transformation)
'log' (logarithmic transformation)
'exp' (exponential transformation)
'quantile' (quantile transformation to transform the time series into a standard Gaussian distribution)
'weighted quantile' (quantile transformation with re-scaling to preserve the variance in the original time series, i.e. results in a Gaussian distribution with mean zero and standard deviation according to the orginal standard deviation)
'nonneg' (Set negative values to zero)
'normalize' (center and standardize the time series)
'standardize' (standardize the time series, i.e. divide by its standard deviation)
'center' (center the time series, i.e. subtract its mean)}
}
\value{
Proxytibble with transformed proxy data in `zoo::zoo` format, or transformed irregular time series object (`zoo::zoo`)
}
\description{
Transform data in an irregular time series object
}
\examples{
# Load Monticchio example data from PTBoxProxydata
library(PTBoxProxydata)
mng <- PTBoxProxydata::ProxyDataManager()
monticchiodata <- PTBoxProxydata::load_set(mng,'monticchio_testset',zoo_format = 'zoo', force_file=TRUE)
# Transform the arboreal pollen signal using logit, square root, and quantile transformations
monticchio_logit <- paleodata_transformation(monticchiodata$proxy_data[[1]]/100,transformation_type="logit")
monticchio_sqrt <- paleodata_transformation(monticchiodata$proxy_data[[1]],transformation_type="sqrt")
monticchio_quantile <- paleodata_transformation(monticchiodata$proxy_data[[1]],transformation_type="quantile")
# Plot zoo data
plot(monticchio_logit)
plot(monticchio_sqrt)
plot(monticchio_quantile)

}
\seealso{
\link{logitlink} (from `VGAM`) for 'logit' and 'invlogit' transformations

\link{probitlink} (from `VGAM`) for 'probit' and 'invprobit' transformations

\link{orderNorm} (from `bestNormalize`) for 'quantile' transformation

\link{normalize} (from `PTBoxProxytools`) for 'normalize', 'standardize', and 'center' transformations
}
