% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_tools.R
\name{paleodata_varfromspec}
\alias{paleodata_varfromspec}
\title{Variance computation by integration of the spectrum, based on PaleoSpec::GetVarFromSpectra()
DEPRICATED --> Use `tsc_and_period_dep_var` instead
This version is a reimplementation based on `PaleoSpec::GetVarFromSpectra`, which is no longer needed}
\usage{
paleodata_varfromspec(
  xin,
  freq.start = NULL,
  freq.end = NULL,
  target = "raw",
  interpolation = TRUE,
  method = "lh14",
  xout = NULL,
  transformation = FALSE,
  transformation_type = "normalize",
  detrend = TRUE,
  df_log = 0.05,
  bLog = FALSE
)
}
\arguments{
\item{xin}{Proxytibble with proxy data in `zoo::zoo` format, or irregular time series object (`zoo::zoo`), xin can be multivariate}

\item{freq.start}{Vector containing the start frequencies of the fitting interval(s)}

\item{freq.end}{Vector containing the start frequencies of the fitting interval(s)}

\item{target}{denotes  "raw" or "logsmooth", default "raw"}

\item{interpolation}{Logical: should interpolation be applied?}

\item{method}{Parameters for interpolation (see paleodata_interpolation)}

\item{xout}{Parameters for interpolation (see paleodata_interpolation). If NULL the interpolation is set to the mean temporal resolution of each proxy time series.}

\item{transformation}{Logical: should transformation be applied?}

\item{transformation_type}{Parameters for transformation (see paleodata_transformation)}

\item{detrend}{Logical: should time series be linearly detrend before computing the spectrum}

\item{df_log}{Width of the log-smoother in log units}

\item{bLog}{TRUE: average in the log space of the power, FALSE: arithmetic average}
}
\value{
List with estimated variance
}
\description{
Variance computation by integration of the spectrum, based on PaleoSpec::GetVarFromSpectra()
DEPRICATED --> Use `tsc_and_period_dep_var` instead
This version is a reimplementation based on `PaleoSpec::GetVarFromSpectra`, which is no longer needed
}
\examples{
# Load ice core example data
library(PTBoxProxydata)
mng <- ProxyDataManager()
icecoredata <- load_set(mng,'icecore_testset',zoo_format = 'zoo')
# compute variance
testout <- icecoredata \%>\% paleodata_varfromspec(.,  target="raw", freq.start = 5e-5, freq.end = 5e-4)
print(testout)

}
\seealso{
\link{GetVarFromSpectra} (from `PaleoSpec`) for regular time series version

\link{paleodata_spectrum} (from `PTBoxProxytools`) for computation of spectra
}
