% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\docType{package}
\name{PTBoxProxytools-package}
\alias{PTBoxProxytools}
\alias{PTBoxProxytools-package}
\title{PTBoxProxytools: PTBox proxy tools for time series processing and analysis}
\description{
Provides tools and helpers for processing and analyzing (paleo) time series data as part of the PaleoToolBox (PTBox)
}
\author{
\strong{Maintainer}: Nils Weitzel \email{nweitzel@iup.uni-heidelberg.de}

Authors:
\itemize{
  \item Beatrice Ellerhoff \email{beatrice.ellerhoff@iup.uni-heidelberg.de}
  \item Moritz Adam \email{madam@iup.uni-heidelberg.de}
}

}
\keyword{internal}
