% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_helpers.R
\name{remove_empty_rows}
\alias{remove_empty_rows}
\title{Remove empty rows (i.e. consisting only of zeros) in a matrix}
\usage{
remove_empty_rows(xin)
}
\arguments{
\item{xin}{Matrix}
}
\value{
Matrix
}
\description{
Remove empty rows (i.e. consisting only of zeros) in a matrix
}
\examples{
testmatrix <- matrix(c(1,0,1,0),ncol=2)
print(testmatrix)
print(remove_empty_rows(testmatrix))

}
