% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_helpers.R
\name{remove_extrapolated_samples}
\alias{remove_extrapolated_samples}
\title{Remove samples in interpolated zoo that are far away from original samples}
\usage{
remove_extrapolated_samples(
  xin_raw,
  xin_interp,
  max_dist = 5 * mean(diff(zoo::index(xin_interp)))
)
}
\arguments{
\item{xin_raw}{Original proxytibble with proxy data in `zoo::zoo` format, or irregular time series object (`zoo::zoo`), or list of zoo::zoo objects (e.g. from proxytibble$proxy_data) xin can be multivariate}

\item{xin_interp}{Zoo with interpolated values (all timeseries from xin_raw need to be interpolated to the same time axis), xin_interp can have dimension 1 (time), 2 (time x records or sites x time), or 3 (sites x time x records)}

\item{max_dist}{Cutoff distance from nearest raw sample beyond which values in xin_interp are set to NA}
}
\value{
zoo of same dimension as xin_interp
}
\description{
Remove samples in interpolated zoo that are far away from original samples
}
