% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_helpers.R
\name{rep_zoo}
\alias{rep_zoo}
\title{Replicate data vector of zoo object (time axis does not change)}
\usage{
rep_zoo(xin, times = 1)
}
\arguments{
\item{xin}{Proxytibble with proxy data in `zoo::zoo` format, or irregular time series object (`zoo::zoo`), xin can be multivariate}

\item{times}{Number of replications}
}
\value{
Proxytibble with proxy data in `zoo::zoo` format, or irregular time series object (`zoo::zoo`), xin can be multivariate
}
\description{
Replicate data vector of zoo object (time axis does not change)
}
\examples{
# Load Monticchio example data from PTBoxProxydata
library(PTBoxProxydata)
mng <- PTBoxProxydata::ProxyDataManager()
monticchiodata <- PTBoxProxydata::load_set(mng,'monticchio_testset',zoo_format = 'zoo', force_file=TRUE)
# Clean proxy data time series
monticchiodata_rep <- rep_zoo(monticchiodata,5)
# Clean zoo time series
monticchiozoo_rep <- rep_zoo(monticchiodata$proxy_data[[1]],5)
# Plot zoo data
plot(monticchiodata_rep$proxy_data[[1]])
plot(monticchiozoo_rep)

}
