% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_helpers.R
\name{resample}
\alias{resample}
\title{Efficient resampling of elements for uncertainty quantification with bootstrapping}
\usage{
resample(x, size = length(x), replace = TRUE, ...)
}
\arguments{
\item{x}{Input data}

\item{size}{Number of output elements, defaults to length(x)}

\item{replace}{Replace entries?, defaults to TRUE}

\item{...}{Other parameters for sample.int}
}
\value{
Shuffled version of input data
}
\description{
Efficient resampling of elements for uncertainty quantification with bootstrapping
}
