% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_helpers.R
\name{rev_time_axis}
\alias{rev_time_axis}
\title{Reverse the time axis of `zoo::zoo`}
\usage{
rev_time_axis(xin)
}
\arguments{
\item{xin}{Proxytibble with proxy data in `zoo::zoo` format, or irregular time series object (`zoo::zoo`), xin can be multivariate}
}
\value{
zoo with returned time axis
}
\description{
Reverse the time axis of `zoo::zoo`
}
\examples{
# Load Monticchio example data from PTBoxProxydata
library(PTBoxProxydata)
mng <- PTBoxProxydata::ProxyDataManager()
monticchiodata <- PTBoxProxydata::load_set(mng,'monticchio_testset',zoo_format = 'zoo', force_file=TRUE)
# Change from yrs BP to yrs AP (i.e. forward instead of backward) time axis for proxytibble
monticchiodata_rev <- rev_time_axis(monticchiodata)
# Change from yrs BP to yrs AP (i.e. forward instead of backward) time axis for zoo
monticchiozoo_rev <- rev_time_axis(monticchiodata$proxy_data[[1]])
# Plot zoo data
plot(monticchiozoo_rev)
plot(monticchiodata_rev$proxy_data[[1]])

}
