% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_helpers.R
\name{sample_var_uncertainty}
\alias{sample_var_uncertainty}
\title{Sample from asymptotic chi-square distribution of variance estimator (using sample variance and dof)}
\usage{
sample_var_uncertainty(nr_samples, var_estimate, dof)
}
\arguments{
\item{nr_samples}{Number of samples}

\item{var_estimate}{Point estimate of the variance}

\item{dof}{Degrees of freedom}
}
\value{
Samples from the chi-square distribution of the variance estimator
}
\description{
Sample from asymptotic chi-square distribution of variance estimator (using sample variance and dof)
}
