% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_tools.R
\name{tsc_and_period_dep_cor}
\alias{tsc_and_period_dep_cor}
\title{Compute period and timescale specific correlation between irregularly spaced time series (using Gaussian kernel correlation if not all indices of the zoo's coincide)}
\usage{
tsc_and_period_dep_cor(xin, yin, period_start, period_end, tsc_low, tsc_up)
}
\arguments{
\item{xin}{Input zoo}

\item{yin}{Input zoo}

\item{period_start}{Start of period of interest}

\item{period_end}{End of period of interest}

\item{tsc_low}{Upper filtering timescale}

\item{tsc_up}{Lower filtering timescale}
}
\value{
Data frame with correlation, p-value, and upper and lower bounds of the confidence interval columns
}
\description{
Compute period and timescale specific correlation between irregularly spaced time series (using Gaussian kernel correlation if not all indices of the zoo's coincide)
}
