% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_tools.R
\name{tsc_and_period_dep_cormat}
\alias{tsc_and_period_dep_cormat}
\title{Compute period and timescale specific correlation matrix for multivariate zoo (based on common samples if some zoo columns contain NAs)}
\usage{
tsc_and_period_dep_cormat(xin, period_start, period_end, tsc_low, tsc_up)
}
\arguments{
\item{xin}{Input multivariate zoo}

\item{period_start}{Start of period of interest}

\item{period_end}{End of period of interest}

\item{tsc_low}{Upper filtering timescale}

\item{tsc_up}{Lower filtering timescale}
}
\value{
List with correlation matrix, p-value matrix, and matrices for upper and lower bounds of the confidence interval
}
\description{
Compute period and timescale specific correlation matrix for multivariate zoo (based on common samples if some zoo columns contain NAs)
}
