% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_tools.R
\name{tsc_and_period_dep_signal_extraction}
\alias{tsc_and_period_dep_signal_extraction}
\title{Wrapper for extracting signals for specific periods and time scales}
\usage{
tsc_and_period_dep_signal_extraction(
  xin,
  period_start,
  period_end,
  filter_type,
  tsc_low,
  tsc_up,
  signal_type
)
}
\arguments{
\item{xin}{Input zoo}

\item{period_start}{Start of period of interest}

\item{period_end}{End of period of interest}

\item{filter_type}{Type of timescale filtering}

\item{tsc_low}{Upper filtering timescale}

\item{tsc_up}{Lower filtering timescale}

\item{signal_type}{Signal type}
}
\value{
Zoo with desired signal
}
\description{
Wrapper for extracting signals for specific periods and time scales
}
