% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_tools.R
\name{tsc_and_period_dep_var}
\alias{tsc_and_period_dep_var}
\title{Wrapper function for estimating period and timescale-dependent variances (based on PTBoxProxytools and nest functions)}
\usage{
tsc_and_period_dep_var(
  xin,
  period_start,
  period_end,
  tsc_low,
  tsc_up,
  detrend = FALSE
)
}
\arguments{
\item{xin}{Input zoo}

\item{period_start}{Start of period of interest}

\item{period_end}{End of period of interest}

\item{tsc_low}{Upper filtering timescale}

\item{tsc_up}{Lower filtering timescale}

\item{detrend}{Should linear detrending be applied before computing the spectra from which the timescales dependent variance is deduced}
}
\value{
Data frame with columns for variance estimate, dof's, and lower and upper bounds of the confidence intervals
}
\description{
Wrapper function for estimating period and timescale-dependent variances (based on PTBoxProxytools and nest functions)
}
