% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_tools.R
\name{tsc_and_period_dep_var_samples}
\alias{tsc_and_period_dep_var_samples}
\title{Wrapper to draw samples of timescale and time period dependent variances}
\usage{
tsc_and_period_dep_var_samples(
  xin,
  period_start,
  period_end,
  tsc_low,
  tsc_up,
  nr_samples,
  detrend = FALSE
)
}
\arguments{
\item{xin}{Input zoo}

\item{period_start}{Start of period of interest}

\item{period_end}{End of period of interest}

\item{tsc_low}{Upper filtering timescale}

\item{tsc_up}{Lower filtering timescale}

\item{nr_samples}{Number of samples}

\item{detrend}{Should linear detrending be applied before computing the spectra from which the timescales dependent variance is deduced}
}
\value{
Vector (univariate zoo) of matrix (multivariate zoo) with variance samples
}
\description{
Wrapper to draw samples of timescale and time period dependent variances
}
