% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_tools.R
\name{site_mean}
\alias{site_mean}
\title{Constructs stacks of records from a proxytibble (list of proxy records)}
\usage{
site_mean(
  xin,
  nr_samples = 1000,
  lon_min = -180,
  lon_max = 180,
  lat_min = -90,
  lat_max = 90,
  time = NULL,
  period_start = min(time),
  period_end = max(time),
  sample_interval = mean(diff(time)),
  use_complete_records_only = FALSE,
  restrict_to_window = FALSE,
  window_tmin = period_start,
  window_tmax = period_end,
  window_minres = 2500,
  window_maxstep = 5000,
  window_minlength = (period_end - period_start)/2,
  window_maxdistfromsample = 3000,
  xin_interp = NULL,
  age_uncertainty = "white_noise",
  age_uncertainty_std = 1000,
  value_uncertainty = "white_noise",
  value_uncertainty_std = 5,
  interpolation_method = "lh14",
  lh14_lowpass = 1.2,
  lh14_length = 5,
  bin_width = sample_interval,
  binning_function = mean,
  loess_span = 0.25,
  gk_antialiasing = TRUE,
  gk_smoothscale = NULL,
  gk_pass = 0.5,
  gam_family = "gaussian",
  gam_smoothscale = NULL,
  bayesgamls_family = "gaussian",
  bayesgamls_smoothscale = NULL,
  bayesgamls_backend = "rstan",
  site_normalization = FALSE,
  bootstrapping = TRUE,
  stacking_method = "site_mean",
  gridbox_size = c(20, 10),
  land_area_only = TRUE,
  weighting_map = NULL,
  dist_exp = 1,
  within_group_method = "avgdist",
  group_weights = NULL,
  group_maps = NULL,
  anomalies = TRUE,
  rescale = FALSE
)
}
\arguments{
\item{xin}{Input proxy data}

\item{nr_samples}{Number of samples}

\item{lon_min}{Lon min}

\item{lon_max}{Lon max}

\item{lat_min}{Lat min}

\item{lat_max}{Lat max}

\item{time}{Time}

\item{period_start}{Start of period}

\item{period_end}{End of period}

\item{sample_interval}{Sample interval}

\item{use_complete_records_only}{Should only records with values at all timesteps be used?}

\item{restrict_to_window}{Should maximal window fulfilling quality criteria be used?}

\item{window_tmin}{Start of period for window}

\item{window_tmax}{End of period for window}

\item{window_minres}{Min resolution}

\item{window_maxstep}{Maximal time step}

\item{window_minlength}{Minimum length of window}

\item{window_maxdistfromsample}{M}

\item{xin_interp}{Interpolated values}

\item{age_uncertainty}{Method for including age uncertainty}

\item{age_uncertainty_std}{Std. for age uncertainty}

\item{value_uncertainty}{Method for including proxy uncertainty}

\item{value_uncertainty_std}{Std. for proxy uncertainty}

\item{interpolation_method}{Interpolation width}

\item{lh14_lowpass}{Scaling factor for the lowpass frequency. Default is 1.2 (same as in PaleoSpec::MakeEquidistant)}

\item{lh14_length}{Scaling factor for the Length of the filter. Default is 5 (same as in PaleoSpec::MakeEquidistant)}

\item{bin_width}{Width of bins if interpolation method is "binning". Defaults to the mean sample resolution (no variable bin sizes are supported at the moment)}

\item{binning_function}{How should values within one bin be averaged? Default is "mean"}

\item{loess_span}{"span" parameter in loess fitting, controls degree of smoothing}

\item{gk_antialiasing}{Should linear interpolation to higher resolution be applied prior to smoothing to avoid aliasing. Default is "TRUE"}

\item{gk_smoothscale}{Smoothing scale of the Gaussian kernel}

\item{gk_pass}{Gain at the smoothing scale. Default is 0.5}

\item{gam_family}{Response distribution family (unlike for bayesgamls, this parameter requires a family function, any family supported by mgcv is valid)}

\item{gam_smoothscale}{Approximate smoothing scale of the GAM model (translated internally to number of basis functions of the fitted splines)}

\item{bayesgamls_family}{Response distribution family (note that unlike gaminterp, here a string with a family name is required not the actually family function, "beta" is the only valid option at the moment); this is because different families have different parameter names to parameterize the gamls model}

\item{bayesgamls_smoothscale}{Approximate smoothing scale of the GAM model (translated internally to number of basis functions of the fitted splines)}

\item{bayesgamls_backend}{Which stan backend should be used ("rstan" or "cmdstanr")}

\item{site_normalization}{Should site values be normalized?}

\item{bootstrapping}{Use bootstrapping over sites?}

\item{stacking_method}{Stacking method}

\item{gridbox_size}{Gridbox size for grid-based interpolation methods}

\item{land_area_only}{Use land areas only for latitudinal weighting}

\item{weighting_map}{Additional map to weight the land points in land area computation (needs to have same dimension as the land map)}

\item{dist_exp}{Exponent in computing weights with avgdist}

\item{within_group_method}{Averaging method within groups}

\item{group_weights}{Group weight}

\item{group_maps}{List of maps to compute group weights}

\item{anomalies}{Should anomalies from time mean be reported for each output sample}

\item{rescale}{Should each output sample be rescaled to unit variance}
}
\value{
List with ensemble mean, ensemble quantiles, and ensemble members of stack
}
\description{
Constructs stacks of records from a proxytibble (list of proxy records)
}
