% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_helpers.R
\name{spatial_means}
\alias{spatial_means}
\title{Calculate spatial mean of a climate data field}
\usage{
spatial_means(lon, lat, clim_field)
}
\arguments{
\item{lon}{Vector of longitudes}

\item{lat}{Vector of latitudes}

\item{clim_field}{Climate data field (with dimension #lon x #lat)}
}
\value{
Number (spatial mean)
}
\description{
Calculate spatial mean of a climate data field
}
\examples{
# Create dummy field
lon <- seq(0,90,by=10)
lat <- seq(0,70,by=10)
clim_field <- matrix(1:80,nrow=10)
# Compute spatial mean
clim_mean <- spatial_means(lon,lat,clim_field)
# Print spatial mean
print(clim_mean)

}
