% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytools_helpers.R
\name{stack_records}
\alias{stack_records}
\title{Creates (weighted) average of proxy timeseries, using selected method for weighting the records}
\usage{
stack_records(
  site_data,
  stacking_method = "site_mean",
  lon_min = -180,
  lon_max = 180,
  lat_min = -90,
  lat_max = 90,
  gridbox_size = c(20, 10),
  land_area_only = TRUE,
  weighting_map = NULL,
  dist_exp = 1,
  within_group_method = "avgdist",
  group_weights = NULL,
  group_maps = NULL
)
}
\arguments{
\item{site_data}{List with site data (lon, lat, var)}

\item{stacking_method}{Stacking method}

\item{lon_min}{Lon min}

\item{lon_max}{Lon max}

\item{lat_min}{Lat min}

\item{lat_max}{Lat max}

\item{gridbox_size}{Gridbox size for grid-based interpolation methods}

\item{land_area_only}{Use land areas only for latitudinal weighting}

\item{weighting_map}{Additional map to weight the land points in land area computation (needs to have same dimension as the land map)}

\item{dist_exp}{Exponent in computing weights with avgdist}

\item{within_group_method}{Averaging method within groups}

\item{group_weights}{Group weight}

\item{group_maps}{List of maps to compute group weights}
}
\value{
Vector wit weighted mean values for each timestep (not a zoo!)
}
\description{
Creates (weighted) average of proxy timeseries, using selected method for weighting the records
}
